+++ docs/src/amsr.jl
@@ -1,6 +1,6 @@
 # North Atlantic Sea Surface Temperature
 using OceanAnalysis, Plots, Dates
-f = get_amsr_file("2025-09-07");
+f = get_amsr("2025-09-07");
 a = read_amsr(f, "SST");
 plot_amsr(a, xlims=(290.0, 340.0), ylims=(30.0, 60.0), color=:turbo,
     levels=0.0:2.5:30.0, clim=(0, 30))